backend/requirements.txt
Django==6.0.1
djangorestframework==3.15.1
django-cors-headers>=4.3
django-mssql-backend
pyodbc

pytest
pytest-django
python-dotenv>=1.0

pytz
sqlparse
asgiref
tzdata

pytest
pytest-django
npm
colorama
typing_extensions
